Change --with-softnet to --net-softnet
`--with-softnet` flag for `targ create` command was deprecated
[a while ago](https://github.com/cirruslabs/tart/releases/tag/0.31.0).
The flag got renamed (temporary aliased in fact) to `--net-softnet`
to match the existing `--net-bridge` flag.

For a while `--with-softnet` was still working, but with the
[0.37.1 release](https://github.com/cirruslabs/tart/releases/tag/0.37.1)
at 2023-02-07 it was totally removed.

Since that moment nesting doesn't work with Tart any more and approaches
to create the VM end with:

```
creating nesting vm: rpc error: code = Unknown desc = starting vm: run nesting-0p7b2911 --no-graphics --with-softnet: exit status 64 (Error: Unknown option '--with-softnet'. Did you mean '--net-softnet'?
Usage: tart run [<options>] <name>
  See 'tart run --help' for more information.
```

This commit updates the flag to use the flag that exists across
multiple versions and is the only one existing in the newer versions.

That, however, means that 0.31.0 becomes the minimum required
version of Tart that nesting may support.